backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
sentence-transformers
torch
torchvision
Pillow
numpy
pgvector
python-multipart
